Kb artickles

+++ modules/td-task/Database/migrations/2024_10_02_400010_create_comments_table.php
@@ -7,13 +7,15 @@ class CreateCommentsTable extends Migration
 {
     public function up()
     {
-        Schema::create('comments', function (Blueprint $table) {
-            $table->id();
-            $table->text('comment');
-            $table->foreignId('task_id')->constrained('tasks')->onDelete('cascade'); // Link to task
-            $table->foreignId('user_id')->constrained('users'); // Link to user who commented
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('comments')) {
+            Schema::create('comments', function (Blueprint $table) {
+                $table->id();
+                $table->text('comment');
+                $table->foreignId('task_id')->constrained('tasks')->onDelete('cascade'); // Link to task
+                $table->foreignId('user_id')->constrained('users'); // Link to user who commented
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
